requirements.txt
wheel
EbookLib
beautifulsoup4
azure-cognitiveservices-speech
just_playback
pdfminer.six
# tkinter